Feature gate TFIDF & TextRank modules
Closes #17
